+++ src/main/java/com/fabiolima/e_commerce/controller/BasketItemController.java
@@ -9,6 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -80,7 +81,7 @@ public class BasketItemController {
     // get the total item price
     @Operation(summary = "Retrieve the total price of an item")
     @GetMapping("/{itemId}/total-price")
-    public ResponseEntity<Double> getTotalItemPrice(@PathVariable("itemId") Long itemId){
+    public ResponseEntity<BigDecimal> getTotalItemPrice(@PathVariable("itemId") Long itemId){
         return ResponseEntity.ok(basketItemService.calculateItemTotalPrice(itemId));
     }
